pyproject.toml
[tool.poetry]
name = "xta_tools"
version = "0.1.0"
description = ""
authors = ["Ko Nakagawa <ko.nakagawa@ffri.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
r2pipe = "^1.5.0"
keystone-engine = "^0.9.2"
typer = "^0.3.2"
r2pyapi = "^0.1.2"
lief = "^0.10.1"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
flake8 = "^3.7.9"
isort = "^5.6.4"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
